fix: FilterInfo short cast and ResourceDictionary InitializeComponent
- Cast FilterInfo position arg to short (API requires short, not int)
- Replace XAML-based ResourceDictionary with programmatic loading
  to avoid InitializeComponent partial class conflict

+++ plugin/NINA.AstroCircular.SkyWaver/SequenceContainers/SkwCollimationRun.cs
@@ -203,7 +203,7 @@ namespace NINA.AstroCircular.SkyWaver.SequenceContainers {
             try {
                 // ── Step 1: Switch filter ──
                 progress?.Report(new ApplicationStatus { Status = $"SKW: Switching to filter {FilterName}" });
-                await filterWheelMediator.ChangeFilter(new FilterInfo(FilterName, 0, 0), ct);
+                await filterWheelMediator.ChangeFilter(new FilterInfo(FilterName, 0, (short)0), ct);
 
                 // ── Step 2: Plate-solve and center on target star (IN FOCUS) ──
                 progress?.Report(new ApplicationStatus {
